Changes applied
