+++ src/Controller/HomeController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Repository\ContractorRepository;
 use App\Repository\ServiceTypeRepository;
+use App\Service\ReviewsService;
 use App\Service\SerializerService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -60,16 +61,18 @@ class HomeController extends AbstractController
      * @param string $service
      * @param ServiceTypeRepository $serviceTypeRepository
      * @param SerializerService $serializer
+     * @param ReviewsService $reviewsService
      * @return JsonResponse
      */
     public function getContractorsByServiceType(
         string $service,
         ServiceTypeRepository $serviceTypeRepository,
-        SerializerService $serializer
+        SerializerService $serializer,
+        ReviewsService $reviewsService
     ): JsonResponse {
         $contractors = $serviceTypeRepository->findOneBy(['name' => $service])->getContractors();
         $json = $serializer->getResponse($contractors, ['filtered']);
-        $json = $serializer->reformatReviews($json);
+        $json = $reviewsService->reformatReviews($json);
 
         return new JsonResponse($json);
     }
